pyproject.toml
[tool.poetry]
name = "m3u8dl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
m3u8 = "^3.2.0"
requests = "^2.28.1"
pycryptodomex = "^3.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
